SecurityConfiguration deprecated methods fix - CorsConfig refactorized and added
