--- src/main/java/com/copypoint/api/infra/http/config/CorsConfig.java
@@ -1,4 +1,0 @@
-package com.copypoint.api.infra.http.config;
-
-public class CorsConfig {
-}
